Create README.md
The Challenge: 
When a client approached us with a request to analyze the results of a game testing session, little did we know that we were in for a wild ride. With over 100 participants testing two versions of the same game, the client hoped to compare the outcomes and identify which version was better received. However, the dataset we received was far from clean.

The Problems:
1.	Timing Issues: The game was supposed to be exactly 5 minutes long, but many participants took longer, either because of technical issues or taking a break.
2.	Complex Requirements: The client wanted to see the average performance of the participants by the second, over a 5-minute period.
3.	Messy Data: The participants' taps were recorded as data points, but with three options - correct, incorrect, or missed - sorting through the data was a nightmare.

The Solution: 
To tackle the messiness of the data, we had to do some heavy-duty data wrangling and manipulation. We combed through the entries to identify and remove any instances where the game went over 5 minutes, either due to technical issues or taking a break. The result was a cleaner dataset that was ready for visualization.

The Results: 
With the data in order, we were able to create a series of stunning visualizations that brought the insights to life. We compared the performance of the two versions of the game, showing how the presence of a timer impacted the participants' taps. We also created a series of graphs that illustrated the average performance of the participants over the 5-minute period, broken down by seconds.
The client was beyond thrilled with the results, as he never expected that the complex requirements could be met with this messy dataset. He now had a clear picture of which version of the game was better received and could make data-driven decisions to improve the game experience.
In conclusion, the challenge of unlocking insights from messy game data was not an easy one, but with some innovative data manipulation and visualization, we were able to turn the data into a powerful tool for the client.